requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
plaid-python==38.4.0
python-dotenv==1.0.1
httpx==0.27.2
pydantic==2.9.2
google-auth==2.29.0
google-auth[requests]
google-auth-oauthlib==1.2.1
google-api-python-client==2.151.0
beautifulsoup4==4.12.3
lxml==5.3.0
PyJWT==2.8.0
requests==2.32.3
psycopg2-binary==2.9.9
python-multipart==0.0.9
